# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salary Manager",
        "description": "CRUD&Permisson+Role system.",
        "contact": {
            "email": "Hexagon14@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/currency": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Get list of currencies.",
                "description": "Returns list of currencies.",
                "operationId": "getCurrencyList",
                "responses": {
                    "200": {
                        "description": "Returns list of currencies."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Creates currency.",
                "description": "Creates currency.",
                "operationId": "createCurrency",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "rate": {
                                        "type": "string"
                                    },
                                    "symbol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "dollar",
                                    "rate": "5",
                                    "symbol": "$"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns currency."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/api/currency/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Get one currency.",
                "description": "Returns one currency.",
                "operationId": "getCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns currency."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Deletes currency.",
                "description": "Deletes currency.",
                "operationId": "deleteCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency deleted."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Updates currency.",
                "description": "Updates currency parameters that were given in request.",
                "operationId": "updateCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "rate": {
                                        "type": "string"
                                    },
                                    "symbol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "dollar",
                                    "rate": "5",
                                    "symbol": "$"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns currency."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Bearer Token": {
                "type": "apiKey",
                "description": "Laravel passport security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/auth/login",
                        "tokenUrl": "http://localhost/auth/token",
                        "refreshUrl": "http://localhostauth/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}